Accept (but log) errors during push_ssh_keys_to_models
This is an attempt to work around issues regarding CAAS models where
the KeyManager class isn't supported, and thus where this code would
otherwise fail.  I couldn't find a clearly documented way to determine
if a model was a CAAS model, and didn't want to depend on undocumented
behavior to make things work for now.