Fix the asyncio issue
Applying Robert Gildein's fix to the asyncio issue, his words:

I could reproduce this only on 3.8.10 Python, on 3.10.4 is not there.

To fix this you need to use model = mock.MagicMock() instead of
model = mock.Mock(). Why? Not sure, libjuju required

   websockets<8.0,>=7.0; python_version < "3.9"

and if I install version 10.3 the issue disappears.

IMO it's ok to ignore that issue.